backend/requirements.txt
fastapi
uvicorn
python-multipart
anthropic
python-dotenv
PyMuPDF
python-docx
reportlab
